+++ src/main/java/com/apex/model/texture/Texture.java
@@ -1,7 +1,7 @@
 package com.apex.model.texture;
 
 public interface Texture {
-    int getPixelColor(float u, float v);
+    int getPixelColor(double u, double v);
 
     String getCache();
 }
